refactor: move files to AudioDec directory and remove test.py
