+++ setup.py
@@ -1,7 +1,7 @@
 from setuptools import setup, find_packages
 
 setup(
-    name='audiodec',
+    name='AudioDec',
     version='1.0',
     packages=find_packages(),
     install_requires=[],
